[#3] SignInViewController에 네비게이션 코드 추가
