+++ Feelter/Sources/Feature/Join/JoinViewController.swift
@@ -1,0 +1,16 @@
+//
+//  JoinViewController.swift
+//  Feelter
+//
+//  Created by 이정동 on 7/30/25.
+//
+
+import UIKit
+
+final class JoinViewController: RxBaseViewController {
+
+    override func viewDidLoad() {
+        view.backgroundColor = .brightTurquoise
+        title = "회원가입"
+    }
+}
